Desktop/run_nsfw/requirements.txt
matplotlib
numpy
pillow
torch
torchvision
tqdm
argparse